refactor(specialists): extract read-only queries to agents/queries.ts
Move the read-only query cluster out of src/lib/agents.ts:
- listRunningAgentsSync
- listAgentStates
- listRunningAgents
- warnOnBareNumericIssueIds
- dropLegacyAgentStatesMissingRoleAsync

Keep agents.ts as the public surface by importing the moved functions
from ./agents/queries.js and re-exporting them. Export isRole so the
new module can share the existing type guard without duplication.

No state writers moved; lint:state-writes stays green.
